Move EC2Instance from slurm_resources to fleet_manager module.
Fleet manager module should not depend on something which is defined in slurm_plugin.
