+++ tests/slurm_plugin/test_clustermgtd.py
@@ -21,13 +21,13 @@ import pytest
 import slurm_plugin
 from assertpy import assert_that
 from slurm_plugin.clustermgtd import ClusterManager, ClustermgtdConfig, ComputeFleetStatus, ComputeFleetStatusManager
+from slurm_plugin.fleet_manager import EC2Instance
 from slurm_plugin.slurm_resources import (
     EC2_HEALTH_STATUS_UNHEALTHY_STATES,
     EC2_INSTANCE_ALIVE_STATES,
     EC2_SCHEDULED_EVENT_CODES,
     ComputeResourceFailureEvent,
     DynamicNode,
-    EC2Instance,
     EC2InstanceHealthState,
     PartitionStatus,
     SlurmPartition,
